refactor(model): Restructure classes to support ludo game logic
